Change parameter encoding, Update models, Add Validation, Add necessary extensions, UI Changes for adding new value, Add new value logic, Update project structure

+++ fit-swift-client/Core/Utility/Extensions.swift
@@ -98,7 +98,7 @@ extension Bundle {
     }
 }
 
-extension Double {
+extension Float {
     var truncateTrailingZeros: String {
         return truncatingRemainder(dividingBy: 1) == 0 ? String(format: "%g", self) : String(self)
     }
